requirements.txt
django
djangorestframework
paho-mqtt
scikit-learn
numpy
pandas
joblib